+++ src/core_modules/capture-core/components/DataEntry/dataEntryOutput/withWarningOutput.js
@@ -1,66 +1,19 @@
 // @flow
 import * as React from 'react';
 import { connect } from 'react-redux';
-import Card from '@material-ui/core/Card';
-import { IconWarningFilled16 } from '@dhis2/ui';
-import { withStyles } from '@material-ui/core/styles';
-import i18n from '@dhis2/d2-i18n';
 import { getDataEntryKey } from '../common/getDataEntryKey';
 import { withDataEntryOutput } from './withDataEntryOutput';
+import { WidgetWarning } from '../../WidgetErrorAndWarning/WidgetWarning';
 
 
 type Props = {
     warningItems: ?Array<any>,
     warningOnCompleteItems: ?Array<any>,
     saveAttempted: boolean,
-    classes: {
-        list: string,
-        listItem: string,
-        card: string,
-        header: string,
-        headerText: string,
-    },
 };
-
-const styles = theme => ({
-    list: {
-        margin: 0,
-    },
-    listItem: {
-        paddingLeft: theme.typography.pxToRem(10),
-        marginTop: theme.typography.pxToRem(8),
-    },
-    header: {
-        display: 'flex',
-        alignItems: 'center',
-    },
-    headerText: {
-        marginLeft: theme.typography.pxToRem(10),
-    },
-    card: {
-        borderRadius: theme.typography.pxToRem(5),
-        padding: theme.typography.pxToRem(10),
-        backgroundColor: theme.palette.warning.lighter,
-    },
-
-});
 
 const getWarningOutput = () =>
     class WarningOutputBuilder extends React.Component<Props> {
-        static renderWarningItems = (warningItems: any, classes: any) =>
-            (<div>
-                {warningItems &&
-                    warningItems.map(item => (
-                        <li
-                            key={item.id}
-                            className={classes.listItem}
-                        >
-                            <p>{item.message}</p>
-                        </li>
-                    ),
-                    )}
-            </div>)
-
         getVisibleWarningItems() {
             const { warningItems, warningOnCompleteItems, saveAttempted } = this.props;
             if (saveAttempted) {
@@ -76,26 +29,8 @@ const getWarningOutput = () =>
         }
 
         render = () => {
-            const { classes } = this.props;
             const visibleItems = this.getVisibleWarningItems();
-            return (
-                <div>
-                    {visibleItems && visibleItems.length > 0 &&
-                    <Card className={classes.card}>
-                        <div className={classes.header}>
-                            <IconWarningFilled16 />
-                            <div className={classes.headerText}>
-                                {i18n.t('Warnings')}
-                            </div>
-                        </div>
-                        <ul className={classes.list}>
-                            {WarningOutputBuilder.renderWarningItems(visibleItems, classes)}
-                        </ul>
-                    </Card>
-                    }
-                </div>
-
-            );
+            return <WidgetWarning warning={visibleItems} />;
         }
     };
 
@@ -117,4 +52,4 @@ export const withWarningOutput = () =>
     (InnerComponent: React.ComponentType<any>) =>
         withDataEntryOutput()(
             InnerComponent,
-            withStyles(styles)(connect(mapStateToProps, mapDispatchToProps)(getWarningOutput())));
+            connect(mapStateToProps, mapDispatchToProps)(getWarningOutput()));