+++ src/core_modules/capture-core/components/FormFields/Options/SingleSelectBoxes/singleSelectBoxes.types.js
@@ -13,7 +13,7 @@ export type Props = {
         label: string,
         iconSelected: string,
         iconDeselected: string,
-        checkbox: string,
+        radio: string,
     },
     style?: ?Object,
 };