+++ src/core_modules/capture-core/components/Pages/EnrollmentEditEvent/EnrollmentEditEventPage.container.js
@@ -247,6 +247,7 @@ const EnrollmentEditEventPageWithContextPlain = ({
             teiId={teiId}
             enrollmentId={enrollmentId}
             eventId={eventId}
+            stageId={stageId}
             trackedEntityTypeId={trackedEntityTypeId}
             enrollmentsAsOptions={enrollmentsAsOptions}
             teiDisplayName={teiDisplayName}